+++ src/Symfony/Component/DependencyInjection/Dumper/PhpDumper.php
@@ -223,11 +223,11 @@ EOF;
             }
             if ($ids = array_keys($ids)) {
                 sort($ids);
-                $c = "<?php\n\nreturn array(\n";
+                $c = "<?php\n\nreturn [\n";
                 foreach ($ids as $id) {
                     $c .= '    '.$this->doExport($id)." => true,\n";
                 }
-                $files['removed-ids.php'] = $c .= ");\n";
+                $files['removed-ids.php'] = $c .= "];\n";
             }
 
             foreach ($this->generateServiceFiles() as $file => $c) {
@@ -266,11 +266,11 @@ if (!\\class_exists({$options['class']}::class, false)) {
     \\class_alias(\\Container{$hash}\\{$options['class']}::class, {$options['class']}::class, false);
 }
 
-return new \\Container{$hash}\\{$options['class']}(array(
+return new \\Container{$hash}\\{$options['class']}([
     'container.build_hash' => '$hash',
     'container.build_id' => '$id',
     'container.build_time' => $time,
-), __DIR__.\\DIRECTORY_SEPARATOR.'Container{$hash}');
+], __DIR__.\\DIRECTORY_SEPARATOR.'Container{$hash}');
 
 EOF;
         } else {
@@ -913,12 +913,12 @@ use Symfony\Component\DependencyInjection\ParameterBag\FrozenParameterBag;
 class $class extends $baseClass
 {
     private \$parameters;
-    private \$targetDirs = array();
+    private \$targetDirs = [];
 
     /*{$this->docStar}
      * @internal but protected for BC on cache:clear
      */
-    protected \$privates = array();
+    protected \$privates = [];
 
     public function __construct()
     {
@@ -936,7 +936,7 @@ EOF;
         }
         if ($this->asFiles) {
             $code = str_replace('$parameters', "\$buildParameters;\n    private \$containerDir;\n    private \$parameters", $code);
-            $code = str_replace('__construct()', '__construct(array $buildParameters = array(), $containerDir = __DIR__)', $code);
+            $code = str_replace('__construct()', '__construct(array $buildParameters = [], $containerDir = __DIR__)', $code);
             $code .= "        \$this->buildParameters = \$buildParameters;\n";
             $code .= "        \$this->containerDir = \$containerDir;\n";
         }
@@ -956,7 +956,7 @@ EOF;
         if ($this->container->getParameterBag()->all()) {
             $code .= "        \$this->parameters = \$this->getDefaultParameters();\n\n";
         }
-        $code .= "        \$this->services = \$this->privates = array();\n";
+        $code .= "        \$this->services = \$this->privates = [];\n";
 
         $code .= $this->addSyntheticIds();
         $code .= $this->addMethodMap();
@@ -968,7 +968,7 @@ EOF;
 
     public function reset()
     {
-        \$this->privates = array();
+        \$this->privates = [];
         parent::reset();
     }
 
@@ -1036,7 +1036,7 @@ EOF;
             }
         }
 
-        return $code ? "        \$this->syntheticIds = array(\n{$code}        );\n" : '';
+        return $code ? "        \$this->syntheticIds = [\n{$code}        ];\n" : '';
     }
 
     private function addRemovedIds(): string
@@ -1063,7 +1063,7 @@ EOF;
                 $code .= '            '.$this->doExport($id)." => true,\n";
             }
 
-            $code = "array(\n{$code}        )";
+            $code = "[\n{$code}        ]";
         }
 
         return <<<EOF
@@ -1087,7 +1087,7 @@ EOF;
             }
         }
 
-        return $code ? "        \$this->methodMap = array(\n{$code}        );\n" : '';
+        return $code ? "        \$this->methodMap = [\n{$code}        ];\n" : '';
     }
 
     private function addFileMap(): string
@@ -1101,16 +1101,16 @@ EOF;
             }
         }
 
-        return $code ? "        \$this->fileMap = array(\n{$code}        );\n" : '';
+        return $code ? "        \$this->fileMap = [\n{$code}        ];\n" : '';
     }
 
     private function addAliases(): string
     {
         if (!$aliases = $this->container->getAliases()) {
-            return "\n        \$this->aliases = array();\n";
+            return "\n        \$this->aliases = [];\n";
         }
 
-        $code = "        \$this->aliases = array(\n";
+        $code = "        \$this->aliases = [\n";
         ksort($aliases);
         foreach ($aliases as $alias => $id) {
             $id = (string) $id;
@@ -1120,7 +1120,7 @@ EOF;
             $code .= '            '.$this->doExport($alias).' => '.$this->doExport($id).",\n";
         }
 
-        return $code."        );\n";
+        return $code."        ];\n";
     }
 
     private function addInlineRequires(): string
@@ -1168,7 +1168,7 @@ EOF;
                 throw new InvalidArgumentException(sprintf('Parameter name cannot use env parameters: %s.', $resolvedKey));
             }
             $export = $this->exportParameters([$value]);
-            $export = explode('0 => ', substr(rtrim($export, " )\n"), 7, -1), 2);
+            $export = explode('0 => ', substr(rtrim($export, " ]\n"), 2, -1), 2);
 
             if (preg_match("/\\\$this->(?:getEnv\('(?:\w++:)*+\w++'\)|targetDirs\[\d++\])/", $export[1])) {
                 $dynamicPhp[$key] = sprintf('%scase %s: $value = %s; break;', $export[0], $this->export($key), $export[1]);
@@ -1176,7 +1176,8 @@ EOF;
                 $php[] = sprintf('%s%s => %s,', $export[0], $this->export($key), $export[1]);
             }
         }
-        $parameters = sprintf("array(\n%s\n%s)", implode("\n", $php), str_repeat(' ', 8));
+
+        $parameters = sprintf("[\n%s\n%s]", implode("\n", $php), str_repeat(' ', 8));
 
         $code = <<<'EOF'
 
@@ -1246,14 +1247,14 @@ EOF;
 EOF;
             $getDynamicParameter = sprintf($getDynamicParameter, implode("\n", $dynamicPhp));
         } else {
-            $loadedDynamicParameters = 'array()';
+            $loadedDynamicParameters = '[]';
             $getDynamicParameter = str_repeat(' ', 8).'throw new InvalidArgumentException(sprintf(\'The dynamic parameter "%s" must be defined.\', $name));';
         }
 
         $code .= <<<EOF
 
     private \$loadedDynamicParameters = {$loadedDynamicParameters};
-    private \$dynamicParameters = array();
+    private \$dynamicParameters = [];
 
     /*{$this->docStar}
      * Computes a dynamic parameter.
@@ -1310,7 +1311,7 @@ EOF;
             $php[] = sprintf('%s%s => %s,', str_repeat(' ', $indent), $this->export($key), $value);
         }
 
-        return sprintf("array(\n%s\n%s)", implode("\n", $php), str_repeat(' ', $indent - 4));
+        return sprintf("[\n%s\n%s]", implode("\n", $php), str_repeat(' ', $indent - 4));
     }
 
     private function endClass(): string
@@ -1404,7 +1405,7 @@ EOF;
                 $code[] = sprintf('%s => %s', $this->dumpValue($k, $interpolate), $this->dumpValue($v, $interpolate));
             }
 
-            return sprintf('array(%s)', implode(', ', $code));
+            return sprintf('[%s]', implode(', ', $code));
         } elseif ($value instanceof ArgumentInterface) {
             $scope = [$this->definitionVariables, $this->referenceVariables];
             $this->definitionVariables = $this->referenceVariables = null;