+++ src/Symfony/Component/HttpFoundation/HeaderUtils.php
@@ -31,7 +31,7 @@ class HeaderUtils
      * Example:
      *
      *     HeaderUtils::split("da, en-gb;q=0.8", ",;")
-     *     // => array(array('da'), array('en-gb', 'q=0.8'))
+     *     // => ['da'], ['en-gb', 'q=0.8']]
      *
      * @param string $header     HTTP header value
      * @param string $separators List of characters to split on, ordered by
@@ -75,8 +75,8 @@ class HeaderUtils
      *
      * Example:
      *
-     *     HeaderUtils::combine(array(array("foo", "abc"), array("bar")))
-     *     // => array("foo" => "abc", "bar" => true)
+     *     HeaderUtils::combine([["foo", "abc"], ["bar"]])
+     *     // => ["foo" => "abc", "bar" => true]
      */
     public static function combine(array $parts): array
     {
@@ -99,7 +99,7 @@ class HeaderUtils
      *
      * Example:
      *
-     *     HeaderUtils::toString(array("foo" => "abc", "bar" => true, "baz" => "a b c"), ",")
+     *     HeaderUtils::toString(["foo" => "abc", "bar" => true, "baz" => "a b c"], ",")
      *     // => 'foo=abc, bar, baz="a b c"'
      */
     public static function toString(array $assoc, string $separator): string