+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -25,7 +25,7 @@ CHANGELOG
  * removed the `Serializer::$normalizerCache` and `Serializer::$denormalizerCache`
    properties
  * added an optional `string $format = null` argument to `AbstractNormalizer::instantiateObject`
- * added an optional `array $context = array()` to `Serializer::supportsNormalization`, `Serializer::supportsDenormalization`,
+ * added an optional `array $context = []` to `Serializer::supportsNormalization`, `Serializer::supportsDenormalization`,
    `Serializer::supportsEncoding` and `Serializer::supportsDecoding`
 
 3.4.0